+++ main.tf
@@ -197,7 +197,7 @@ resource "aws_cloudfront_distribution" "this" {
 
   viewer_certificate {
     ssl_support_method       = "sni-only"
-    minimum_protocol_version = "TLSv1.3_2025"
+    minimum_protocol_version = var.cloudfront_min_proto_version
     acm_certificate_arn      = aws_acm_certificate_validation.this.certificate_arn
   }